Change SingletonMeta base to ABCMeta.
This will allow any class that has ABCMeta as a metaclass to also be a singleton. Derived classes metas i.e. SingletonMeta muse be a subclass of the meta of all bases. Hence any class that has ABCMeta rather than type as a metaclass will not be able to use singleton unless it too is a subclass of ABCMeta.

First noticed and fixed for my config package.

	modified:   singleton/_singleton.py
	modified:   tests/test_singleton.py